Add generic pattern replacer and profiling
This brings the generic pattern replacer interface for pattern/replace
based optimizations.
This also brings the zero pattern replacement, which is exposed as
the "zero" optimization parameter ("-O zero" flag).

This also brings changes to allow in-program profiling. This is nothing
serious yet. It simply prints out the program state before terminating.